Updated frontend to use one websocket
Combined laser state and robot coordinates into one object sent between back and front-end

Refactored TypeScript so that the websockets are handled in a standalone class which is called by the main ts file

When sending a state change, it will be based off of the current value of the elements in the front-end. I added a function where it will re-render the UI if necessary when it gets a state change.